create useItemRoot hook
